examples/google-adk-agent/requirements.txt
google-adk>=0.3.0
